remove utils::token mod
