+++ sway-lsp/src/capabilities/semantic_tokens.rs
@@ -1,8 +1,7 @@
 use crate::core::{
     session::Session,
-    token::{SymbolKind, Token},
+    token::{get_range_from_span, SymbolKind, Token},
 };
-use crate::utils::token::get_range_from_span;
 use std::sync::{
     atomic::{AtomicU32, Ordering},
     Arc,
