+++ sway-lsp/src/utils/debug.rs
@@ -1,6 +1,5 @@
 #![allow(dead_code)]
-use crate::core::token::Token;
-use crate::utils::token::get_range_from_span;
+use crate::core::token::{get_range_from_span, Token};
 use sway_core::language::Literal;
 use sway_types::{Ident, Spanned};
 use tower_lsp::lsp_types::{Diagnostic, DiagnosticSeverity};
